Eliminating some excess printing

+++ firstRunInit.m
@@ -2,6 +2,7 @@ function firstRunInit()
     %% If the scripts folder does not exist, create it, along with all the scripts inside it
     folder = '/home/eilam.morag/hw_injection/Hardware_Injection_2016/scripts';
     if (~exist(folder, 'dir'))
+        disp('First time initializations...');
         disp(['Creating the folder: ', folder]);
         mkdir(folder);
         % Add the functions for creating the runAll scripts
@@ -31,22 +32,26 @@ function firstRunInit()
     clear folder pulsar
     %% If the main webpage does not exist, create it
     homepage = '/home/eilam.morag/public_html/HWInjection/HWInjection.php';
-%     if (~exist(homepage, 'file'))
+    if (~exist(homepage, 'file'))
         disp(['Creating the homepage: ', homepage]);
         homepageHTML();
         homepageCSS();
-%     end
+    end
     %% Copy the JavaScript files to the public_html/HWInjection directory
     destination = '/home/eilam.morag/public_html/HWInjection/';
     file1 = 'changeDate.js';
     file2 = 'readableDate.js';
-    status1 = copyfile(file1, destination);
-    status2 = copyfile(file2, destination);
-    if (status1 == 1)
-        disp(['Copying the file: ', file1]);
+    path1 = [destination, file1];
+    path2 = [destination, file2];
+    if (~exist(path1, 'file') || ~exist(path2, 'file'))
+        status1 = copyfile(file1, destination);
+        status2 = copyfile(file2, destination);
+        if (status1 == 1)
+            disp(['Copying the file: ', file1]);
+        end
+        if (status2 == 1)
+            disp(['Copying the file: ', file2]);
+        end
+        clear destination file1 file2 status1 status2 
     end
-    if (status2 == 1)
-        disp(['Copying the file: ', file2]);
-    end
-    clear destination file1 file2 status1 status2 
 end
